move simple querying to its own class
